setup socket.io and server/client instances
