+++ .gitignore
@@ -1,1 +1,4 @@
 /frontend/node_modules/*
+/user-service/node_modules/*
+/matching-service/client/node_modules/*
+/matching-service/server/node_modules/*
