+++ matching-service/client/client.js
@@ -1,0 +1,3 @@
+import { io } from "socket.io-client";
+
+const socket = io('http://localhost:8001')
