+++ matching-service/client/package.json
@@ -1,0 +1,12 @@
+{
+  "type": "module",
+  "scripts": {
+    "dev": "nodemon client.js",
+    "start": "node client.js",
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "dependencies": {
+    "nodemon": "^2.0.19",
+    "socket.io-client": "^4.5.1"
+  }
+}
