--- matching-service/package.json
@@ -1,25 +1,0 @@
-{
-  "name": "matching-service",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "type": "module",
-  "scripts": {
-    "dev": "nodemon index.js",
-    "start": "node index.js",
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "devDependencies": {
-    "nodemon": "^2.0.19"
-  },
-  "dependencies": {
-    "express": "^4.18.1",
-    "sequelize": "^6.21.3",
-    "socket.io": "^4.5.1",
-    "sqlite3": "^5.0.10",
-    "uuid": "^8.3.2"
-  }
-}
