+++ user-service/.env
@@ -1,0 +1,2 @@
+ENV=PROD
+DB_CLOUD_URI=mongodb+srv://cs3219grp49:dockerCompose@cluster0.flgp7sz.mongodb.net/?retryWrites=true&w=majority
